% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ciplot}
\alias{ciplot}
\alias{plot.cuminc2}
\alias{lines.cuminc2}
\title{\code{cuminc2} plotting method}
\usage{
ciplot(
  x,
  col.ci = seq_along(xx),
  lty.ci = par("lty"),
  lwd.ci = par("lwd"),
  conf.int = FALSE,
  alpha.conf = 0.05,
  col.conf = NULL,
  lty.conf = 2L,
  lwd.conf = lwd.ci,
  conf.band = FALSE,
  alpha.band = 0.5,
  events = TRUE,
  atrisk = TRUE,
  events.total = TRUE,
  wh.events = c("events", "est", "est.sd", "est.ci", "atrisk", "percent", "percent.ci"),
  events.lab = NULL,
  events.pad = 0.5,
  events.digits = ifelse(grepl("percent", wh.events), 0L, 3L),
  events.lines = TRUE,
  events.col = FALSE,
  include_censored = FALSE,
  main = NULL,
  xlab = "Time",
  ylab = "Probability",
  groups.lab = names(xx),
  xlim = NULL,
  ylim = NULL,
  cex.lab = par("cex.lab"),
  cex.main = par("cex.main"),
  cex.axis = par("cex.axis"),
  cex.events = cex.axis,
  gy_test = TRUE,
  test_details = TRUE,
  legend.args = list(),
  split = FALSE,
  xaxis.at = pretty(xlim),
  yaxis.at = pretty(ylim),
  xaxis.lab = xaxis.at,
  yaxis.lab = yaxis.at,
  events.at = xaxis.at,
  groups.order = seq_along(xx),
  extra.margin = 5L,
  mar = NULL,
  add = FALSE,
  panel.first = NULL,
  panel.last = NULL,
  ...
)

\method{plot}{cuminc2}(
  x,
  col.ci = seq_along(xx),
  lty.ci = par("lty"),
  lwd.ci = par("lwd"),
  conf.int = FALSE,
  alpha.conf = 0.05,
  col.conf = NULL,
  lty.conf = 2L,
  lwd.conf = lwd.ci,
  conf.band = FALSE,
  alpha.band = 0.5,
  events = TRUE,
  atrisk = TRUE,
  events.total = TRUE,
  wh.events = c("events", "est", "est.sd", "est.ci", "atrisk", "percent", "percent.ci"),
  events.lab = NULL,
  events.pad = 0.5,
  events.digits = ifelse(grepl("percent", wh.events), 0L, 3L),
  events.lines = TRUE,
  events.col = FALSE,
  include_censored = FALSE,
  main = NULL,
  xlab = "Time",
  ylab = "Probability",
  groups.lab = names(xx),
  xlim = NULL,
  ylim = NULL,
  cex.lab = par("cex.lab"),
  cex.main = par("cex.main"),
  cex.axis = par("cex.axis"),
  cex.events = cex.axis,
  gy_test = TRUE,
  test_details = TRUE,
  legend.args = list(),
  split = FALSE,
  xaxis.at = pretty(xlim),
  yaxis.at = pretty(ylim),
  xaxis.lab = xaxis.at,
  yaxis.lab = yaxis.at,
  events.at = xaxis.at,
  groups.order = seq_along(xx),
  extra.margin = 5L,
  mar = NULL,
  add = FALSE,
  panel.first = NULL,
  panel.last = NULL,
  ...
)

\method{lines}{cuminc2}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{cuminc2}}}

\item{col.ci, lty.ci, lwd.ci}{line color, type, and width for each curve}

\item{conf.int}{logical; if \code{TRUE}, adds a confidence interval for
each curve}

\item{alpha.conf}{the confidence level for \code{conf.int}}

\item{col.conf, lty.conf, lwd.conf}{line color, type, and width for each
confidence interval line}

\item{conf.band}{logical; if \code{TRUE}, a band is draw for confidence
intervals}

\item{alpha.band}{alpha transparency (in \code{[0, 1]}) of the band}

\item{events}{logical; if \code{TRUE}, a cumulative events table is drawn}

\item{atrisk}{logical; if \code{TRUE}, the number at risk is added to the
events table}

\item{events.total}{logical or numeric; if \code{TRUE}, cumulative events
for each group is added to events table at a calculated position; for more
control, use a specific x-coordinate}

\item{wh.events}{a character string giving the type of \code{events} table
to show; one of \code{"events"} (cumulative number of events), \code{"est"}
(estimates, see \code{\link[cmprsk]{timepoints}}), \code{"est.sd"}
(estimate +/- standard deviation), \code{"est.ci"} (estimate with
confidence interval), \code{"atrisk"} (event at-risk table with
censored), \code{"percent"} (percent), or \code{"percent.ci"} (pecent with
confidence interval)}

\item{events.lab}{heading for events table}

\item{events.pad}{extra padding between plot and events table; alternatively,
a vector of padding for each line in the events table, recycled as needed}

\item{events.digits}{when estimates are shown in events table (see
\code{wh.events}), number of digits past the decimal to show}

\item{events.lines}{logical; draw lines next to groups in events table}

\item{events.col}{logical or a vector with colors for events table text;
if \code{TRUE}, \code{col.ci} will be used}

\item{include_censored}{logical; if \code{TRUE}, censored patients are
included in at-risk counts (may not be desired); the default (\code{FALSE})
will only sum the at-risk table rows}

\item{main}{title of plot}

\item{xlab, ylab}{x- and y-axis labels}

\item{groups.lab}{labels for each line in \code{events} table}

\item{xlim, ylim}{x- and y-axis limits}

\item{cex.lab, cex.main}{text size for axis and main titles}

\item{cex.axis}{text size for axis labels and \code{gy_test}events table}

\item{cex.events}{text size for events table}

\item{gy_test}{logical; if \code{TRUE} the tests of group equality will
be shown}

\item{test_details}{logical; if \code{TRUE} (default), all test details
(test statistic, degrees of freedom, p-value) are shown; if \code{FALSE},
only the p-value is shown}

\item{legend.args}{an optional \emph{named} list of \code{\link{legend}}
arguments controlling the annotations when \code{gy_test = TRUE}}

\item{split}{optionally split plot by unique competing risks or group;
one of \code{FALSE} (default, no splitting), \code{"group"}, or
\code{"event"}}

\item{xaxis.at, yaxis.at}{positions for x- and y-axis labels and ticks}

\item{xaxis.lab, yaxis.lab}{x- and y-axis tick labels}

\item{events.at}{x-coordinates to show events table (default is
\code{xaxis.at})}

\item{groups.order}{order of groups in events table}

\item{extra.margin}{increase left margin when groups labels in events
table are long (note that this will be overridden by \code{mar})}

\item{mar}{margins; see \code{mar} section in \code{\link{par}}}

\item{add}{logical; if \code{TRUE}, \code{par}s are not reset; allows for
multiple panels, e.g., when using \code{par(mfrow = c(1, 2))}}

\item{panel.first}{an expression to be evaluated after the plot axes are
set up but before any plotting takes place}

\item{panel.last}{an expression to be evaluated after plotting but before
returning from the function}

\item{...}{additional parameters (\code{font}, \code{mfrow}, \code{bty},
\code{tcl}, \code{cex.lab}, \code{xaxs}, etc) passed to \code{\link{par}}}
}
\description{
Plot a \code{\link{cuminc2}} object with optional cumulative events or
estimate table, Gray's test results, and other features.
}
\examples{
tp <- within(transplant, {
  futime <- futime + 1e-8
  age50 <- factor(+(age > 50))
  age_cat <- cut(age, c(0, 40, 60, Inf), c('<40', '40-60', '60+'))
})

## basic usage
ci1 <- cuminc2(Surv(futime, event(censored)) ~ 1, tp)
plot(ci1)
plot(ci1, groups.lab = paste('1', names(ci1$cuminc))) ## testing
plot(ci1, conf.int = TRUE, lty.conf = 3)
plot(ci1, conf.band = TRUE, add = TRUE)
lines(ci1, col = c(NA, 'blue', NA), lwd = 3)

ci1 <- cuminc2(Surv(futime, event(censored)) ~ age_cat, tp)
plot(ci1)
plot(ci1, split = 'event')
plot(ci1, split = 'event', events = FALSE, test_details = FALSE,
     legend.args = list(x = 'topright', cex = 1.5, text.col = 2,
                        title = "Gray\'s test p-value for"))

## also plots "cuminc" objects but without extra features
plot(ci1$cuminc)

ci1 <- cuminc2(Surv(futime, event(censored) == death) ~ age50, tp)
plot(ci1, lty.ci = c(1, 1, 2, 2, 3, 3), col.ci = 1:2)

ci2 <- cuminc2(Surv(futime, event(censored) == death) ~ 1, tp)
plot(ci2, wh.events = 'est', events.digits = 2, groups.order = c(2, 1, 3))

}
\seealso{
\code{\link{ciplot_by}}; \code{\link{cuminc2}};
\code{\link{summary.cuminc2}}; \code{\link[cmprsk]{cuminc}}
}
