% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuminc2.R
\name{ciplot_by}
\alias{ciplot_by}
\title{ciplot_by}
\usage{
ciplot_by(rhs = "1", event, data, by = NULL, single = TRUE,
  cencode = NULL, gy_test = TRUE, main = NULL, ylab = NULL,
  sub = NULL, strata_lab = NULL, fig_lab = NULL, col.ci = NULL,
  map.col = FALSE, time = NULL, add = FALSE, plot = TRUE, ...)
}
\arguments{
\item{rhs}{the right-hand side of the formula}

\item{event}{character string indicating the event; see details}

\item{data}{data frame to use}

\item{by}{optional character string of stratification variable}

\item{single}{logical; if TRUE, each level of by will be drawn in a separate window}

\item{cencode}{unique value of \code{event} denoting censoring}
}
\description{
This function helps create stratified \code{\link{ciplot}}s quickly with
panel labels and Gray's tests for each plot.
}
\examples{
## basic usage: Surv(time, event) ~ 1
ciplot_by(time = 'futime', event = 'event', data = transplant)

## with groups: Surv(time, event) ~ group
ciplot_by('sex', time = 'futime', event = 'event',
          data = transplant)
ciplot_by('sex', time = 'futime', event = 'event',
          data = transplant, by = 'abo', single = FALSE)

}
