% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{coef.crr}
\alias{coef.crr}
\alias{coefficients.crr}
\title{ct \code{crr} model coefficients}
\usage{
\method{coef}{crr}(object, ...)

\method{coefficients}{crr}(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link[cmprsk]{crr}}}

\item{...}{ignored}
}
\description{
ct \code{crr} model coefficients
}

