% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{crrfit}
\alias{AIC.crr}
\alias{BIC.crr}
\alias{crrfit}
\alias{deviance.crr}
\alias{extractAIC.crr}
\alias{extractBIC.crr}
\alias{extractIC}
\alias{logLik.crr}
\title{\code{crr} model fit statistics}
\usage{
\method{extractAIC}{crr}(object, ...)

\method{AIC}{crr}(object, ...)

extractBIC.crr(object, ...)

\method{BIC}{crr}(object, ...)

extractIC(object, ic = c("AIC", "BIC", "AICc", "BICc"), p)

\method{logLik}{crr}(object, ...)

\method{deviance}{crr}(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link[cmprsk]{crr}}}

\item{...}{ignored}

\item{ic}{information criterion, one of \code{"AIC"}, \code{"BIC"},
\code{"AICc"}, or \code{"BICc"}; see details}

\item{k}{numeric specifying the "weight" of the equivalent degrees of
freedom (\code{=: edf}) part in the AIC formula}
}
\description{
Functions to assess the quality of fitted \code{\link[cmprsk]{crr}} models.

Extract AIC from a \code{crr} model

Select multiple types of Akaike or Bayesian (Schwarz) information
criterion from a \code{\link[cmprsk]{crr}} object to assess the relative
quality of models for a given data set.
}
\details{
\code{AIC} and \code{BIC} are calculated in the usual way. \code{AICc}
is the AIC with a correction for finite sample sizes. This assumes that
the model is univariate, linear, and has normally-distributed residuals
(conditional upon regressors).

\code{BICc} is a newly proposed criteria that is a modified BIC for
competing risks data subject to right censoring (Kuk, 2013).
}
\examples{
fit <- crr2(Surv(futime, event(censored) == death) ~ age, transplant)
extractAIC(fit[[1]])
AIC(fit[[1]])


}
\references{
Kuk D, Varadhan R. Model selection in competing risks regression.
\emph{Stat Med}. 2013 Aug \strong{15};32.

Kuk D, Varadhan R. Model selection in competing risks regression.
\emph{Stat Med}. 2013 Aug \strong{15};32.
}

