% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{crrwald.test}
\alias{crrwald.test}
\title{Wald test for \code{crr} model coefficients}
\usage{
crrwald.test(object, terms)
}
\arguments{
\item{object}{an object of class \code{\link[cmprsk]{crr}}}

\item{terms}{a list (optionally named) giving the indices of the
coefficients to test; a list of length \code{n} will perform \code{n}
tests; default is an overall test and each coefficient individually

Alternatively, a (named) vector (or list) of character strings can be
used which will \code{\link{grep}} for each pattern in the coefficient
names; patterns that match multiple terms will be tested as a group, and
only one character string per test will be used}
}
\description{
Wald test for \code{crr} model coefficients
}
\examples{
crrs <- crr2(Surv(futime, event(censored) == death) ~
               age + sex + abo, transplant)

## default is to test all estimates overall and individually
crrwald.test(crrs[[1L]])
summary(crrs[[1]])$coef[, 'p-value']


## identical ways to call crrwald.test
crrwald.test(crrs[[1L]], list(overall = 1:5, Age = 1,
                              Sex = 2, ABO = 3:5))

crrwald.test(crrs[[1L]], c(overall = '.', Age = 'age', ## or list()
                           Sex = 'sex', ABO = 'abo'))

}
\seealso{
\code{\link[aod]{wald.test}}
}

