% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuminc2.R
\name{cuminc_pairs}
\alias{cuminc_pairs}
\title{Pairwise \code{cuminc} comparisons}
\usage{
cuminc_pairs(object, data = NULL, rho = 0, cencode = NULL,
  method = p.adjust.methods, digits = 3L)
}
\arguments{
\item{object}{a \code{\link{cuminc2}} formula or object of class
\code{\link{cuminc2}}}

\item{data}{a data frame to use (ignored if \code{formula}) is a
\code{\link{cuminc2}} object}

\item{rho, cencode}{passed to \code{\link[cmprsk]{cuminc}}}

\item{method}{p-value correction method (default is \code{'holm'};
see \code{\link{p.adjust}}}

\item{digits}{integer indicating the number of decimal places to be used}
}
\description{
Evaluate pairwise group differences in cumulative incidence with
\code{\link[cmprsk]{cuminc}}.
}
\examples{
## these two are equivalent since there are only two levels
form <- Surv(futime, event(censored)) ~ sex
cuminc_pairs(form, transplant)
cuminc2(form, transplant)$cuminc$Tests


## more useful in the case of three or more levels
form <- Surv(futime, event(censored)) ~ abo
cuminc_pairs(form, transplant)$p.value
cuminc2(form, transplant)$cuminc$Tests


tp <- transplant[transplant$abo \%in\% c('A', 'B', 'O'), ]
tp <- tp[tp$event \%in\% c('censored', 'ltx', 'death'), ]
ci <- cuminc2(form, tp)
ci[[1L]]$Tests
ciplot(ci)

cuminc_pairs(ci)$p.value$ltx

}
