% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crr2.R
\name{finegray2}
\alias{finegray2}
\title{\code{finegray2}}
\usage{
finegray2(formula, data, cencode, ...)
}
\arguments{
\item{formula}{a standard model formula, with survival on the left and
covariates on the right}

\item{data}{a data frame containing the variables in \code{formula}}

\item{cencode}{optional censor code; if missing, the censor code will be
guessed from the status variable}

\item{...}{additional arguments passed to \code{\link{finegray}} or
\code{\link{coxph}}}
}
\description{
Fit multiple \code{\link[survival]{finegray}} models.
}
\examples{
## example from survival::finegray
mg <- within(mgus2, {
  etime <- ifelse(pstat == 0, futime, ptime)
  event <- ifelse(pstat == 0, 2 * death, 1)
  event <- factor(event, 0:2, c('censor', 'pcm', 'death'))
})

pdata <- finegray(Surv(etime, event) ~ age + sex, data = mg)
coxph(Surv(fgstart, fgstop, fgstatus) ~ age + sex,
      weight = fgwt, data = pdata)

fg <- finegray2(Surv(etime, event) ~ age + sex, data = mg)
fg$pcm

## use the call and data to evaluate the equivalent coxph model
eval(fg$pcm$call, list(fgdata = fg$pcm$fgdata))

}
