% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{parse_formula}
\alias{parse_formula}
\title{Parse \code{cmprsk2} formula}
\usage{
parse_formula(formula, data = NULL, name = NULL)
}
\arguments{
\item{formula}{a valid \code{crr2} or \code{cuminc2} formula}

\item{data, name}{optional data frame and data name to check variables
from \code{formula}}
}
\value{
A list with the following elements:

\item{\code{formula}}{the unchanged \code{formula} coerced to a
  \code{\link{formula}} object}
\item{\code{lhs_vars}}{all variable names found on the left-hand side of
  \code{formula}}
\item{\code{rhs_vars}}{all variable names found on the right-hand side of
  \code{formula} with any \code{factor()}, \code{strata()}, etc.
  modifications removed}
\item{\code{lhs}}{the left-hand side of \code{formula} as it appears}
\item{\code{rhs}}{the right-hand side of \code{formula} as it appears with
  any \code{factor()}, \code{strata()}, etc. included}
\item{\code{ftime}}{variable in \code{data} defining times}
\item{\code{fstatus}}{variable in \code{data} defining censoring and
  failure events}
\item{\code{cencode}}{censor code, a value of \code{data$fstatus}}
\item{\code{failcode}}{failure code of interest if given, a value of
  \code{[data$fstatus]} (required for a \code{crr2} model but optional for
  \code{cuminc2})}
\item{\code{cov1}}{\code{cov1} variable names to be evaluated by
  \code{\link{model.matrix}} as they appear}
\item{\code{cov2}}{\code{cov2} variable names (functions of time) to be
  evaluated by \code{\link{model.matrix}} as they appear}
\item{\code{strata}}{strata variable names}
}
\description{
Parse a \code{\link{crr2}} or \code{\link{cuminc2}} formula into useful
pieces and optionally check data source for errors.
}
