% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crr2.R
\name{print.crr2}
\alias{print.crr2}
\title{\code{crr2} print method}
\usage{
\method{print}{crr2}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{crr2}}}

\item{...}{ignored}
}
\description{
\code{crr2} print method
}
\seealso{
\code{\link{crr2}}; \code{\link{summary.crr2}}
}
