% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuminc2.R
\name{print.cuminc2}
\alias{print.cuminc2}
\title{\code{cuminc2} print method}
\usage{
\method{print}{cuminc2}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{cuminc2}}}

\item{...}{ignored}
}
\description{
\code{cuminc2} print method
}
\seealso{
\code{\link{cuminc2}}; \code{\link{summary.cuminc2}}
}
