% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{summary.crr2}
\alias{summary.crr2}
\title{\code{crr2} summary method}
\usage{
\method{summary}{crr2}(object, conf.int = 0.95, ..., html = TRUE,
  combine_ci = FALSE, digits = 2L, format_p = TRUE)
}
\arguments{
\item{object}{an object of class \code{"crr2"}}

\item{...}{additional arguments affecting the summary produced}

\item{html}{logical; if \code{TRUE}, an \code{\link{htmlTable}} will be
returned; otherwise, a matrix}

\item{combine_ci}{logical; if \code{FALSE}, upper and lower confidence
limits will be returned as separate columsn; otherwise, an interval string
will be returned}

\item{digits}{integer value indicating number of digits to print}

\item{format_p}{logical; if \code{TRUE}, p-values will be formatted;
otherwise, p-values will only be rounded}
}
\description{
\code{crr2} summary method
}

