% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuminc2.R
\name{summary.cuminc2}
\alias{summary.cuminc2}
\title{\code{cuminc2} summary method}
\usage{
\method{summary}{cuminc2}(object, times = NULL, digits = 5L, ...)
}
\arguments{
\item{object}{an object of class \code{\link{cuminc2}}}

\item{times}{a vector of times}

\item{digits}{number of digits past the decimal point to keep}

\item{...}{ignored}
}
\value{
A list with the following components:

\item{\code{est}}{a matrix of estimates of the subdistributions for the
  cumulative incidence model(s); see \code{\link[cmprsk]{timepoints}}}
\item{\code{var}}{a matrix giving the corresponding variances for the
  cumulative incidence model(s); see \code{\link[cmprsk]{timepoints}}}
\item{\code{events}}{a matrix with the number of cumulative events up
  to and including each value in \code{times}}
\item{\code{total_events}}{a vector giving the total number of events of
  each type (excluding censored observations)}
\item{\code{total_groups}}{a vector giving the number by group (including
  censored observations)}
\item{\code{total_atrisk}}{a vector giving the number remaining in the
  risk set regardless of event or censoring}
}
\description{
\code{cuminc2} summary method
}
\examples{
ci <- cuminc2(Surv(futime, event(censored)) ~ sex, transplant)
summary(ci)

ci <- cuminc2(Surv(futime, event(censored)) ~ 1, transplant)
summary(ci, times = 0:10 * 100)

}
\seealso{
\code{\link{cuminc2}}; \code{\link{print.cuminc2}};
\code{\link[cmprsk]{timepoints}}; \code{\link{timepoints2}}
}
