% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{terms.crr2}
\alias{terms.crr}
\alias{terms.crr2}
\title{Model terms}
\usage{
\method{terms}{crr2}(x, ...)

\method{terms}{crr}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link[cmprsk]{crr}}}

\item{...}{ignored}
}
\description{
Model terms
}

