% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuminc.R
\name{timepoints2}
\alias{print.timepoints2}
\alias{timepoints2}
\title{\code{timepoints} formatter}
\usage{
timepoints2(w, times, digits = 3L, ci = TRUE, html = TRUE)

\method{print}{timepoints2}(x, ..., html = attr(x, "html"))
}
\arguments{
\item{w, times}{arguments passed to \code{\link{timepoints}}}

\item{digits}{number of digits past the decimal point to keep}

\item{ci}{logical; not implemented}

\item{html}{logical; if \code{TRUE}, html-friendly formatted is returned;
the print method for \code{timepoints2} will use \code{\link{htmlTable}}
if \code{html = TRUE}}

\item{x}{an object of class \code{"timepoints2"}}

\item{...}{additional arguments passed to or from other methods}
}
\description{
Format output from \code{\link[cmprsk]{timepoints}} into a single matrix
with estimate +/- standard deviation.
}
\examples{
## example from cmprsk::cuminc
set.seed(2)
ss <- rexp(100)
gg <- factor(sample(1:3,100,replace=TRUE),1:3,c('a','b','c'))
cc <- sample(0:2,100,replace=TRUE)
strt <- sample(1:2,100,replace=TRUE)
print(xx <- cuminc(ss,cc,gg,strt))

timepoints(xx, times = 0:4)
timepoints2(xx, html = FALSE)

print(timepoints2(xx),
      rgroup = c('One', 'Two'), n.rgroup = c(3,3),
      caption = 'Timepoints<sup>&dagger;</sup>',
      tfoot = '<sup>&dagger;</sup>Est &pm; sd')

}

