% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuminc2.R
\name{ciplot}
\alias{ciplot}
\alias{plot.cuminc2}
\title{\code{cuminc2} plotting method}
\usage{
ciplot(x, lty.ci = par("lty"), lwd.ci = par("lwd"),
  col.ci = seq_along(xx), events = TRUE, wh.events = c("events", "est",
  "est.sd", "est.ci"), col.events = col.ci, events.lab = NULL,
  events.digits = 3L, events.lines = TRUE, events.col = FALSE,
  main = NULL, xlab = "Time", ylab = "Probability",
  groups.lab = names(xx), xlim = NULL, ylim = NULL,
  cex.axis = par("cex.axis"), gy_test = TRUE, split = FALSE,
  xaxis.at = pretty(xlim), yaxis.at = pretty(ylim), xaxis.lab = xaxis.at,
  yaxis.lab = yaxis.at, events.at = xaxis.at,
  groups.order = seq_along(xx), extra.margin = 5L, panel.first = NULL,
  panel.last = NULL, add = FALSE, mar = NULL, ...)

\method{plot}{cuminc2}(x, lty.ci = par("lty"), lwd.ci = par("lwd"),
  col.ci = seq_along(xx), events = TRUE, wh.events = c("events", "est",
  "est.sd", "est.ci"), col.events = col.ci, events.lab = NULL,
  events.digits = 3L, events.lines = TRUE, events.col = FALSE,
  main = NULL, xlab = "Time", ylab = "Probability",
  groups.lab = names(xx), xlim = NULL, ylim = NULL,
  cex.axis = par("cex.axis"), gy_test = TRUE, split = FALSE,
  xaxis.at = pretty(xlim), yaxis.at = pretty(ylim), xaxis.lab = xaxis.at,
  yaxis.lab = yaxis.at, events.at = xaxis.at,
  groups.order = seq_along(xx), extra.margin = 5L, panel.first = NULL,
  panel.last = NULL, add = FALSE, mar = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{\link{cuminc2}}}

\item{main}{title of plot}

\item{xlab, ylab}{x- and y-axis labels}

\item{...}{additional parameters (\code{font}, \code{mfrow}, \code{bty},
\code{tcl}, \code{cex.lab}, \code{xaxs}, etc) passed to \code{\link{par}}}
}
\description{
Plot a \code{\link{cuminc2}} object optional cumulative events or estimate
table, Gray's test results, and other features.
}
\examples{
tp <- within(transplant, {
  age50 <- factor(+(age > 50))
  age_cat <- cut(age, c(0, 40, 60, Inf), c('<40', '40-60', '60+'))
})

ci1 <- cuminc2(Surv(futime, event(censored)) ~ age_cat, tp)
plot(ci1)
plot(ci1$cuminc)
plot(ci1, split = 'event')

ci1 <- cuminc2(Surv(futime, event(censored) == death) ~ age50, tp)
plot(ci1, lty.ci = c(1,1,2,2,3,3), col.ci = 1:2)

ci2 <- cuminc2(Surv(futime, event(censored) == death) ~ 1, tp)
plot(ci2, wh.events = 'est')

}
\seealso{
\code{\link{cuminc2}}; \code{\link{summary.cuminc2}}
}
