% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{crr2}
\alias{crr2}
\alias{print.crr2}
\title{Competing risks regression}
\usage{
crr2(formula, data, which = NULL, cox = FALSE, ...)

\method{print}{crr2}(x, ...)
}
\arguments{
\item{formula}{a formula object of form
\code{Surv(time, status(censor) == failure) ~ response} where \code{censor}
and \code{failure} are labels of the \code{status} variable denoting
the censor and failure of interest; note all other unique values of
\code{status} will be considered competing risks}

\item{data}{a data.frame in which to interpret the variables named in
\code{formula}}

\item{which}{optional character string giving the desired outcome of
interest; if given, no other models are returned}

\item{cox}{logical; if \code{TRUE}, a \code{\link{coxph}} model is also
fit modeling the event of interest}

\item{...}{additional arguments passed to \code{\link[cmprsk]{crr}}}
}
\description{
Regression modeling of subdistribution functions in competing risks.
}
\examples{
crr2(Surv(futime, event(censored) == death) ~ age + sex + abo, transplant)

form <- Surv(futime, event(censored) == death) ~ age + sex + abo
crrs <- crr2(form, transplant, which = 'ltx', cox = TRUE)

summary(crrs)

library('htmlTable')
summary(crrs)

crr2(form, transplant, cox = TRUE, variance = FALSE)


## example from cmprsk::crr

# simulated data to test
set.seed(10)
ftime <- rexp(200)
fstatus <- sample(0:2,200,replace=TRUE)
cov <- matrix(runif(600),nrow=200)
dimnames(cov)[[2]] <- c('x1','x2','x3')
print(z <- crr(ftime,fstatus,cov))
summary(z)
# deviance(z)

dat <- data.frame(ftime = ftime, fstatus = fstatus, cov)
(z2 <- crr2(Surv(ftime, fstatus(0) == 1) ~ ., dat, cox = FALSE))
z$call <- z2[[1]]$call <- NULL
stopifnot(identical(z, z2[[1]]))

}
\seealso{
\code{\link{summary.crr2}}; \code{\link[cmprsk]{crr}};
\code{\link[survival]{coxph}}
}

