% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{crrfit}
\alias{crrfit}
\alias{extractIC}
\alias{extractAIC.crr}
\alias{AIC.crr}
\alias{BIC.crr}
\alias{logLik.crr}
\alias{deviance.crr}
\title{\code{crr} model fit statistics}
\usage{
extractIC(object, ic = c("AIC", "BIC", "AICc", "BICc"), p)

\method{extractAIC}{crr}(fit, scale, k = 2, ...)

\method{AIC}{crr}(object, ...)

\method{BIC}{crr}(object, ...)

\method{logLik}{crr}(object, ...)

\method{deviance}{crr}(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link[cmprsk]{crr}}}

\item{ic}{information criterion, one of \code{"AIC"}, \code{"BIC"},
\code{"AICc"}, or \code{"BICc"}; see details}

\item{p}{an optional penalty term to be multiplied by, \code{k}, the
number of free parameters estimated in each model including the intercept
term}

\item{fit, scale, k}{see \code{\link{extractAIC}}}

\item{...}{additional arguments passed to or from other methods}
}
\description{
Functions to assess the quality of fitted \code{\link[cmprsk]{crr}}
and \code{\link{crr2}} objects.

Select multiple types of Akaike or Bayesian (Schwarz) information
criterion from a \code{\link[cmprsk]{crr}} object to assess the relative
quality of models for a given data set.
}
\details{
\code{AIC} and \code{BIC} are calculated in the usual way. \code{AICc}
is the AIC with a correction for finite sample sizes. This assumes that
the model is univariate, linear, and has normally-distributed residuals
(conditional upon regressors).

\code{BICc} is a newly proposed criteria that is a modified BIC for
competing risks data subject to right censoring (Kuk, 2013).
}
\examples{
crrs <- crr2(Surv(futime, event(censored) == death) ~ age, transplant)
crr1 <- crrs[[1L]]

crr1$coef
coef(crr1)
coefficients(crr1)

extractAIC(crr1)
sapply(crrs, AIC)

## these are equivalent
extractIC(crrs[[1L]], p = 0)
-2 * logLik(crrs[[1]])[1]
-2 * crrs[[1]]$loglik

deviance(crr1)

crrFits(crr1)
crrwald.test(crr1)

}
\references{
Kuk D, Varadhan R. Model selection in competing risks regression.
\emph{Stat Med}. 2013 Aug \strong{15};32.
}
\seealso{
\code{\link{crrFits}}; \code{\link{crrwald.test}}
}
