% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{crrFits}
\alias{crrFits}
\title{\code{crr} model selection table}
\usage{
crrFits(..., p)
}
\arguments{
\item{...}{one or more objects of class \code{\link[cmprsk]{crr}}}

\item{p}{a penalty term to be multiplied by, \code{k}, the number of free
parameters estimated in each model including the intercept term}
}
\description{
Return several types of fit statistics for a \code{\link[cmprsk]{crr}}
saturated model compared to the null model. Note that comparisons among
models only make sense for ones fit to the same data set.
}
\examples{
bmt <- read.csv('http://www.stat.unipg.it/luca/R/bmtcrr.csv')
cov1 <- with(bmt, model.matrix(~ Age + relevel(Sex, 'M') + D +
             relevel(Phase, 'Relapse') + Source)[, -1])
m1 <- with(bmt, crr(ftime, Status, cov1))
summary(m1)
aod::wald.test(m1$var, m1$coef, grep('Phase', names(m1$coef)))$result$chi2

## model selection
m2 <- with(bmt, crr(ftime, Status, cov1[, c(4:6)]))
m3 <- with(bmt, crr(ftime, Status, cov1[, c(4:6,7)]))
m4 <- with(bmt, crr(ftime, Status, cov1[, c(4:6,7,1)]))
m5 <- with(bmt, crr(ftime, Status, cov1[, c(4:6,7,2)]))
m6 <- with(bmt, crr(ftime, Status, cov1[, c(4:6,7,3)]))

crrFits(m1, m2, m3, m4, m5, m6, p = 3)

par(mfrow = c(1,3))
for (ii in 1:ncol(m2$res))
  scatter.smooth(m2$uftime, m2$res[, ii],
                 main = names(coef(m2))[ii],
                 xlab = 'Failure time',
                 ylab = 'Schoenfeld residuals')

pred <- with(bmt, model.matrix(~ levels(relevel(Phase, 'Relapse')))[, -1])
pred <- predict(m2, pred)
plot(pred, xlab = 'Failure time', ylab = 'CIF', lty = 1:4)
legend('topleft', lty = 1:4, title = 'Phase', bty = 'n',
       legend = levels(relevel(bmt$Phase, "Relapse")))


}
\references{
Scrucca L, Santucci A, Aversa F (2009). Regression Modeling of Competing
Risk Using R: An In Depth Guide for Clinicians. \emph{Bone Marrow
Transplantation} (2010) \strong{45}, 1388–1395.
}

