% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crr2.R
\name{summary.crr2}
\alias{summary.crr2}
\title{\code{crr2} summary method}
\usage{
\method{summary}{crr2}(object, conf.int = 0.95, ..., html = FALSE,
  combine_ci = FALSE, digits = 2L, format_p = html)
}
\arguments{
\item{object}{an object of class \code{\link{crr2}}}

\item{conf.int}{level of confidence}

\item{...}{additional arguments affecting the summary produced (passed
to \code{\link[cmprsk]{summary.crr}} or
\code{\link[survival]{summary.coxph}} depending on the object)}

\item{html}{logical; if \code{TRUE}, an \code{\link{htmlTable}} will be
returned; otherwise, a matrix}

\item{combine_ci}{logical; if \code{FALSE}, upper and lower confidence
limits will be returned as separate columsn; otherwise, an interval string
will be returned}

\item{digits}{integer value indicating number of digits to print}

\item{format_p}{logical; if \code{TRUE}, p-values will be formatted;
otherwise, p-values will only be rounded}
}
\description{
\code{crr2} summary method
}
\examples{
crrs <- crr2(Surv(futime, event(censored) == death) ~ age + sex + abo,
             data = transplant)
summary(crrs)

summary(crrs, conf.int = 0.9, digits = 3L,
        combine_ci = TRUE, format_p = TRUE)


library('htmlTable')
summary(crrs, html = TRUE, combine_ci = TRUE)

}
\seealso{
\code{\link{crr2}}; \code{\link{print.crr2}}
}
