% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crr2.R
\name{summary.crr2}
\alias{summary.crr2}
\title{\code{crr2} summary method}
\usage{
\method{summary}{crr2}(
  object,
  conf.int = 0.95,
  n = FALSE,
  ref = FALSE,
  html = FALSE,
  combine_ci = html,
  digits = 2L,
  format_p = html,
  color_p = html,
  format_n = n,
  htmlArgs = list(),
  ...
)
}
\arguments{
\item{object}{an object of class \code{\link{crr2}}}

\item{conf.int}{the level for a two-sided confidence interval on the
coeficients; default is 0.95}

\item{n}{logical; if \code{TRUE}, the sample size and number of events
for each variable are added to the summary}

\item{ref}{logical; if \code{TRUE}, rows with reference categories are
added to results}

\item{html}{logical; if \code{TRUE}, an \code{\link{htmlTable}} will be
returned; otherwise, a matrix}

\item{combine_ci}{logical; if \code{FALSE}, upper and lower confidence
limits will be returned as separate columns; otherwise, an interval string
will be returned}

\item{digits}{number of digits past the decimal point to keep}

\item{format_p}{logical; if \code{TRUE}, p-values will be formatted;
otherwise, p-values will only be rounded}

\item{color_p}{logical; if \code{TRUE}, p-values will be formatted and
colored based on significance level; see \code{cmprsk2:::color_pval}}

\item{format_n}{logical; if \code{TRUE}, for \code{html = TRUE} the total
n is added for each total/events column and percents of total and events
are shown in each row}

\item{htmlArgs}{for \code{html = TRUE}, a \emph{named} list of arguments
passed to \code{\link[htmlTable]{htmlTable}} for additional formatting or
to override defaults}

\item{...}{ignored}
}
\description{
\code{crr2} summary method
}
\examples{
crrs <- crr2(Surv(futime, event(censored) == death) ~ age + sex + abo,
             data = transplant)
summary(crrs)

summary(crrs, conf.int = 0.9, digits = 3L,
        combine_ci = TRUE, format_p = TRUE)

library('htmlTable')
summary(crrs, html = TRUE, combine_ci = TRUE, n = TRUE)

summary(
  crrs, html = TRUE, combine_ci = TRUE, n = TRUE, ref = TRUE,
  htmlArgs = list(
    caption = 'CRR models.', rgroup = c('Age', 'Sex', 'Blood type'),
    rnames = c('+1 year change', 'Female', 'B', 'AB', 'O')
  )
)

}
\seealso{
\code{\link{crr2}}; \code{\link{print.crr2}}
}
