% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuminc2.R
\name{summary.cuminc2}
\alias{summary.cuminc2}
\title{\code{cuminc2} summary method}
\usage{
\method{summary}{cuminc2}(object, times = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{\link{cuminc2}}}

\item{times}{a vector of times}

\item{...}{ignored}
}
\value{
A list with the following components:

\item{\code{est}}{a matrix of estimates of the subdistributions for the
cumulative incidence model(s); see \code{\link[cmprsk]{timepoints}}}
\item{\code{var}}{a matrix giving the corresponding variances for the
cumulative incidence model(s); see \code{\link[cmprsk]{timepoints}}}
\item{\code{events}}{a matrix with the number of cumulative events up
to and including each value in \code{times}}
\item{\code{total_events}}{a matrix giving the total number of events}
}
\description{
\code{cuminc2} summary method
}
\examples{
ci <- cuminc2(Surv(futime, event(censored)) ~ sex + strata(abo),
              data = transplant)
summary(ci)
summary(ci, times = c(0, 250, 500))

}
\seealso{
\code{\link{cuminc2}}; \code{\link{print.cuminc2}};
\code{\link[cmprsk]{timepoints}}; \code{\link{timepoints2}}
}
