% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{terms.crr2}
\alias{terms.crr2}
\alias{terms.crr}
\title{\code{crr2} model terms}
\usage{
\method{terms}{crr2}(x, ...)

\method{terms}{crr}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{crr2}}}

\item{...}{ignored}
}
\description{
\code{crr2} model terms
}
