% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuminc2.R
\name{timepoints2}
\alias{timepoints2}
\title{\code{timepoints} formatter}
\usage{
timepoints2(
  w,
  times = NULL,
  digits = 3L,
  sd = FALSE,
  ci = FALSE,
  html = FALSE,
  htmlArgs = list(),
  ...
)
}
\arguments{
\item{w, times}{arguments passed to \code{\link{timepoints}}}

\item{digits}{number of digits past the decimal point to keep}

\item{sd}{logical; if \code{FALSE}, the standard deviation will not be
shown with the estimate}

\item{ci}{logical; not implemented}

\item{html}{logical; if \code{TRUE}, an html-friendly format is returned;
the print method for \code{timepoints2} will use \code{\link{htmlTable}}
if \code{html = TRUE}}

\item{htmlArgs}{for \code{html = TRUE}, a \emph{named} list of arguments
passed to \code{\link[htmlTable]{htmlTable}} for additional formatting or
to override defaults}

\item{...}{ignored}
}
\description{
Format output from \code{\link[cmprsk]{timepoints}} into a single matrix
with estimate +/- standard deviation.
}
\examples{
ci <- cuminc2(Surv(futime, event(censored)) ~ sex, transplant)
timepoints2(ci, sd = TRUE)
timepoints2(ci, html = TRUE)

print(
  timepoints2(ci, sd = TRUE, html = TRUE),
  rgroup = c('Death', 'Ltx', 'Withdraw'), n.rgroup = c(2, 2, 2),
  caption = 'Timepoints<sup>&dagger;</sup>',
  tfoot = '<sup>&dagger;</sup>Estimate &pm; SD'
)


## example from cmprsk::cuminc
set.seed(2)
ss <- rexp(100)
gg <- factor(sample(1:3, 100, replace = TRUE), 1:3, c('a', 'b', 'c'))
cc <- sample(0:2, 100, replace = TRUE)
strt <- sample(1:2, 100, replace = TRUE)
xx <- cuminc(ss, cc, gg, strt)

timepoints(xx, times = 0:4)
timepoints2(xx, times = 0:4)

}
